added book rating system; deleted unnecessary files

+++ src/main/java/ru/itis/models/UsersBooks.java
@@ -28,7 +28,7 @@ public class UsersBooks {
     private Book book;
 
     @Column(name = "users_book_rating")
-    private Double bookRatingByUser;
+    private Integer bookRatingByUser;
 
     @Enumerated(value = EnumType.STRING)
     @Column(name = "book_status")
